ServerErrorResolved
Issue was with having a copy of ViewAllAddressServlet in two places.
(This issue wasn't readily visible because the server error persisted through multiple projects)

Link for viewAllAddressServlet now semi-working; however  link is currently always re-directing to index.html.

This happens when showAllItems.isEmpty is true. Will troubleshoot through line of a few things that can cause this later and figure this out.